Add initial docs, plus some refactoring

+++ qualipy/cli.py
@@ -3,29 +3,167 @@ import json
 
 import click
 from qualipy.anomaly.anomaly import _run_anomaly
-from qualipy.web.deploy import FlaskDeploy, GUnicornDeploy
-from qualipy.web._config import _Config
 from qualipy.backends.pandas_backend.generator import BackendPandas
 from qualipy.project import create_qualipy_folder, Project
 from qualipy.reports.anomaly import AnomalyReport
 from qualipy.reports.comparison import ComparisonReport
 
 
-DEPLOYMENT_OPTIONS = {"flask": FlaskDeploy, "gunicorn": GUnicornDeploy}
+# DEPLOYMENT_OPTIONS = {"flask": FlaskDeploy, "gunicorn": GUnicornDeploy}
 
 BACKEND = BackendPandas
 
 
 @click.group()
 def qualipy():
+    """
+    The main entrypoint for interacting with Qualipy.
+
+    Note: Nearly all these functions rely on the configuration specification
+
+    """
     pass
 
 
 @qualipy.command()
 @click.argument("config_dir_path")
-@click.option("--db_url", default=None)
-def generate_config(config_dir_path, db_url):
-    create_qualipy_folder(config_dir=config_dir_path, db_url=db_url)
+def generate_config(config_dir_path):
+    """
+    Arguments:
+        config_dir_path: The path to the stored directory
+
+    This will generate a folder with the following subfolders and files:
+        * config.json - The most important file. This controls everything - from how to generate
+            the reports, what anomaly model to use, where the data is stored, etc... See (link here)
+            for documentation on the config
+        * models - This folder will store binary versions of your anomaly models used for all projects
+            specified in the config
+    """
+    create_qualipy_folder(config_dir=config_dir_path)
+
+
+@qualipy.command()
+@click.option(
+    "--project_name",
+    default=None,
+    help="Name of the project. Must correspond to an existing project",
+)
+@click.option(
+    "--config_dir", default=None, help="Name of an existing configuration directory"
+)
+@click.option(
+    "--retrain",
+    default=False,
+    type=bool,
+    help="If set to true, will clear all models and stored anomaly data, and retrain each model",
+)
+def run_anomaly(project_name, config_dir, retrain):
+    """ 
+    Runs the anomaly models for a specific project, based on the config
+    """
+    _run_anomaly(project_name, config_dir, retrain)
+
+
+# @qualipy.command()
+# @click.option("--config_dir", default=None)
+# @click.option("--retrain", default=False, type=bool)
+# def schedule_anomaly(config_dir, retrain):
+#     with open(os.path.join(config_dir, "config.json"), "rb") as cf:
+#         config = json.load(cf)
+
+#     scheduler_conf = config["ANOMALY_SCHEDULER"]
+#     project_names = scheduler_conf["PROJECTS"]
+#     for project_name in project_names:
+#         _run_anomaly(project_name, config_dir, retrain)
+
+
+@qualipy.command()
+@click.argument("config_dir", default=None)
+@click.argument("project_name", default=None, type=str)
+@click.option(
+    "--run_anomaly",
+    default=False,
+    type=bool,
+    help="If set to True, qualipy will first run each anomaly model on the data",
+)
+@click.option(
+    "--clear_anomaly",
+    default=False,
+    type=bool,
+    help="If set to True, qualipy will clear all stored anomalies and retrain each model",
+)
+@click.option(
+    "--only_show_anomaly",
+    default=False,
+    type=bool,
+    help="If set to True, only trends containing an anomaly will be shown in the report",
+)
+@click.option(
+    "--t1",
+    default=None,
+    type=str,
+    help="If set, all visualizations in the report will be starting on this date",
+)
+@click.option(
+    "--t2",
+    default=None,
+    type=str,
+    help="If set, all visualizations in the report will be prior to this date",
+)
+@click.option("--out_file", default=None, type=str, help="Location of the output")
+def produce_anomaly_report(
+    config_dir,
+    project_name,
+    run_anomaly,
+    clear_anomaly,
+    only_show_anomaly,
+    t1,
+    t2,
+    out_file,
+):
+    """
+    CONFIG_DIR: The path to the stored directory
+    
+    PROJECT_NAME: Name of the project you want to run an anomaly report on
+    """
+    view = AnomalyReport(
+        config_dir=config_dir,
+        project_name=project_name,
+        run_anomaly=run_anomaly,
+        retrain_anomaly=clear_anomaly,
+        only_show_anomaly=only_show_anomaly,
+        t1=t1,
+        t2=t2,
+    )
+    rendered_page = view.render(
+        template=f"anomaly.j2", title="Anomaly Report", project_name=project_name
+    )
+    if out_file is None:
+        out_file = f"anomaly_report_{project_name}.html"
+    rendered_page.dump(out_file)
+
+
+@qualipy.command()
+@click.argument("config_dir", default=None)
+@click.argument("comparison_name", default=None, type=str)
+@click.option("--out_file", default=None, type=str, help="The name of the output file")
+def produce_comparison_report(
+    config_dir, comparison_name, out_file,
+):
+    """
+    CONFIG_DIR: The path to the stored directory
+    
+    COMPARISON_NAME: Name of the as specified in the configuration file
+    """
+    view = ComparisonReport(config_dir=config_dir, comparison_name=comparison_name,)
+    rendered_page = view.render(
+        template=f"comparison.j2",
+        title="Comparison Report",
+        project_name=comparison_name,
+    )
+    if out_file is None:
+        out_file = f"comparison_report_{comparison_name}.html"
+    rendered_page.dump(out_file)
 
 
 @qualipy.command()
@@ -34,8 +172,14 @@ def generate_config(config_dir_path, db_url):
 @click.option("--recreate", default=True, type=bool)
 def clear_data(config_dir, project_name, recreate):
     project = Project(config_dir=config_dir, project_name=project_name, re_init=True)
-    project.delete_data(recreate=recreate)
-    project.delete_from_project_config()
+    print(
+        f"Preparing to delete table {project_name} as specified in config dir {config_dir}"
+    )
+    if click.confirm(
+        "Do you wish to continue? Warning - this will permanently delete data"
+    ):
+        project.delete_data(recreate=recreate)
+        project.delete_from_project_config()
 
 
 # @qualipy.command()
@@ -65,84 +209,8 @@ def clear_data(config_dir, project_name, recreate):
 #     )
 #     deployer.run()
 
-
-@qualipy.command()
-@click.option("--project_name", default=None)
-@click.option("--config_dir", default=None)
-@click.option("--retrain", default=False, type=bool)
-def run_anomaly(project_name, config_dir, retrain):
-    _run_anomaly(project_name, config_dir, retrain)
-
-
-@qualipy.command()
-@click.option("--config_dir", default=None)
-@click.option("--retrain", default=False, type=bool)
-def schedule_anomaly(config_dir, retrain):
-    with open(os.path.join(config_dir, "config.json"), "rb") as cf:
-        config = json.load(cf)
-
-    scheduler_conf = config["ANOMALY_SCHEDULER"]
-    project_names = scheduler_conf["PROJECTS"]
-    for project_name in project_names:
-        _run_anomaly(project_name, config_dir, retrain)
-
-
-@qualipy.command()
-@click.argument("config_dir", default=None)
-@click.argument("project_name", default=None, type=str)
-@click.option("--run_anomaly", default=False, type=bool)
-@click.option("--clear_anomaly", default=False, type=bool)
-@click.option("--only_show_anomaly", default=False, type=bool)
-@click.option("--t1", default=None, type=str)
-@click.option("--t2", default=None, type=str)
-@click.option("--out_file", default=None, type=str)
-def produce_anomaly_report(
-    config_dir,
-    project_name,
-    run_anomaly,
-    clear_anomaly,
-    only_show_anomaly,
-    t1,
-    t2,
-    out_file,
-):
-    view = AnomalyReport(
-        config_dir=config_dir,
-        project_name=project_name,
-        run_anomaly=run_anomaly,
-        retrain_anomaly=clear_anomaly,
-        only_show_anomaly=only_show_anomaly,
-        t1=t1,
-        t2=t2,
-    )
-    rendered_page = view.render(
-        template=f"anomaly.j2", title="Anomaly Report", project_name=project_name
-    )
-    if out_file is None:
-        out_file = f"anomaly_report_{project_name}.html"
-    rendered_page.dump(out_file)
-
-
-@qualipy.command()
-@click.option("--config_dir", default=None)
-@click.option("--comparison_name", default=None, type=str)
-@click.option("--out_file", default=None, type=str)
-def produce_comparison_report(
-    config_dir, comparison_name, out_file,
-):
-    view = ComparisonReport(config_dir=config_dir, comparison_name=comparison_name,)
-    rendered_page = view.render(
-        template=f"comparison.j2",
-        title="Comparison Report",
-        project_name=comparison_name,
-    )
-    if out_file is None:
-        out_file = f"comparison_report_{comparison_name}.html"
-    rendered_page.dump(out_file)
-
-
 if __name__ == "__main__":
-    # I do the following the debug cli commands, ignore
+    # I do the following to debug cli commands, ignore
 
     import sys
